+++ src/App.js
@@ -1,30 +1,28 @@
-import { useState } from "react";
-import Banner from "./components/Banner/Banner.jsx";
-import Main from './components/Main/Main.jsx';
-import { NavigationContext } from "./store/navigation-context.jsx";
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import Root from "./Root.js";
+import Company from "./components/Main/Company/Company.jsx";
+import Metrics, { thresholdsLoader } from "./components/Main/Metrics/Metrics.jsx";
+import Selector, { filtersLoader } from "./components/Main/Selector/Selector.jsx";
+import SelectorContextProvider from "./store/selector-context.jsx";
+import User from "./components/Main/User/User.jsx";
 
 import "./App.css";
 
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <Root />,
+    children: [
+      { index: true, element: <Metrics />, loader: thresholdsLoader },
+      { path: '/selector', element: <SelectorContextProvider><Selector /></SelectorContextProvider>, loader: filtersLoader},
+      { path: '/company', element: <Company />},
+      { path: '/user', element: <User />},
+    ],
+  },
+]);
+
 function App() {
-  const [selectedNavItem, setSelectedNavItem] = useState("Metrics");
-
-  function selectNavItemHandler(selectedItem) {
-    setSelectedNavItem(() => selectedItem);
-  }
-
-  const ctxNavigation = {
-    selectedNavItem: selectedNavItem,
-    selectNavItemHandler: selectNavItemHandler,
-  };
-
-  return (
-    <>
-      <NavigationContext.Provider value={ctxNavigation}>
-        <Banner />
-        <Main />
-      </NavigationContext.Provider>      
-    </>
-  );
+  return <RouterProvider router={router} />;
 }
 
 export default App;