+++ src/components/Main/Selector/SelectorForm/Button/Button.jsx
@@ -1,0 +1,11 @@
+import { $Button } from "./Button.styles.jsx";
+
+export default function Button({ enabled, name, children }) {
+
+    return(
+        <$Button $enabled={enabled}>
+            <div>{ children }</div>
+            &nbsp;{ name }
+        </ $Button>
+    )
+}