+++ src/components/Main/Selector/SelectorForm/Select/SelectValue.jsx
@@ -1,0 +1,44 @@
+import { useContext, useEffect, useState } from "react";
+import { SelectorContext } from "../../../../../store/selector-context.jsx";
+import $Select from "./Select.styles.jsx";
+import { apiCallToGetFilterValues } from "../../../../../api/Api.jsx";
+
+export default function SelectValue() {
+  const [filterValues, setFilterValues] = useState([]);
+  const [firstOption, setFirstOption] = useState("");
+  const { tempWhereCriteria, setTempWhereCriteria } =
+    useContext(SelectorContext);
+
+  useEffect(() => {
+    async function fetchData() {
+      try {
+        await apiCallToGetFilterValues(tempWhereCriteria.key).then(
+          (resData) => {
+            setFilterValues((_prevFilterValues) => resData);
+            setTempWhereCriteria({
+              ...tempWhereCriteria,
+              value: resData ? resData[0].name : null,
+            });
+            setFirstOption(resData ? resData[0].name : "");
+          }
+        );
+      } catch (error) {
+        //TODO
+      }
+    }
+    fetchData();
+  }, [tempWhereCriteria.key]);
+
+  function handleWhereCriteriaChange(event) {
+    setFirstOption(event.target.value);
+    setTempWhereCriteria({ ...tempWhereCriteria, value: event.target.value });
+  }
+
+  return (
+    <$Select onChange={handleWhereCriteriaChange} value={firstOption} disabled={filterValues.length === 0}>
+      {filterValues.map((filterValue) => (
+        <option key={filterValue.id}>{filterValue.name}</option>
+      ))}
+    </$Select>
+  );
+}