+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+0.21 (2024-07-27)
+===============
+* [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/11)
+* Install react-router-dom and implement router
+* Add Metrics and Selector nav items routing
+* Add loaders for Metrics and Selector routes
+* Create and implement context and reducer for selector
+* Implement new api call for filters and response handler
+* Create selector form
+
 0.20 (2024-06-25)
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/10)